moví app.js fuera de src
